refactor: create new Masonry-renderLoadingItems visual test doc
