+++ packages/gestalt/src/Masonry.tsx
@@ -62,7 +62,7 @@ type Props<T> = {
       | null
       | undefined,
   ) => void;
-  loadingStateItems: ReadonlyArray<T>;
+  loadingStateItems?: ReadonlyArray<T>;
   /**
    * Masonry internally caches item heights using a measurement store. If `measurementStore` is provided, Masonry will use it as its cache and will keep it updated with future measurements. This is often used to prevent re-measurement when users navigate away from and back to a grid. Create a new measurement store with `Masonry.createMeasurementStore()`.
    */
@@ -85,7 +85,7 @@ type Props<T> = {
     readonly itemIdx: number;
     readonly isMeasuring: boolean;
   }) => ReactNode;
-  renderLoadingItems: (arg1: {
+  renderLoadingItems?: (arg1: {
     readonly data: T;
     readonly itemIdx: number;
     readonly isMeasuring: boolean;
